+++ lib/weneedfeed.rb
@@ -5,6 +5,7 @@ require 'weneedfeed/version'
 module Weneedfeed
   autoload :Application, 'weneedfeed/application'
   autoload :Capture, 'weneedfeed/capture'
+  autoload :Command, 'weneedfeed/command'
   autoload :Controllers, 'weneedfeed/controllers'
   autoload :Item, 'weneedfeed/item'
   autoload :Page, 'weneedfeed/page'