+++ lib/weneedfeed/views/show_feed.rb
@@ -11,6 +11,11 @@ module Weneedfeed
 
       private
 
+      # @return [String]
+      def top_page_path
+        request.path.delete_suffix(router.path(:feed, page_name: page_name))
+      end
+
       # @return [Enumerable<Weneedfeed::Item>]
       def items
         @page.items.sort_by do |item|
@@ -21,6 +26,11 @@ module Weneedfeed
       # @return [String]
       def page_name
         request.env['router.params'][:page_name]
+      end
+
+      # @return [Hanami::Router]
+      def router
+        ::Weneedfeed::Application.router
       end
     end
   end